fix(NumberRequestHandler): 이동 시도 횟수를 저장하는 변수의 데이터 타입을 integer에서 long으로 변경
- long으로 변경하여 더 큰 수도 처리할 수 있도록 함
- BigInteger를 사용하면 연산이 느려지는 등 성능이 저하되므로 사용을 지양

+++ src/main/java/racingcar/controller/RacingController.java
@@ -21,7 +21,7 @@ public class RacingController {
 
     public void run() {
         List<String> names = nameRequestHandler.getNames();
-        int moveNum = numberRequestHandler.getNumber();
+        long moveNum = numberRequestHandler.getNumber();
         Cars cars = Cars.from(names);
 
         cars.racing(moveNum);
